refactor: Move out forall tests into forall.rs
